+++ userland/include/api/syscalls.h
@@ -51,7 +51,7 @@ void* mmap(void* addr, size_t length, int prot, int flags, int fd, off_t offset)
 int munmmap(void* addr, size_t length);
 int mprotect(void* addr, size_t length, int prot);
 void* sys_mremap(void* old_address, size_t old_size, size_t new_size, int flags, void* new_address);
-int fork(void);
+uint64_t fork(void);
 
 #ifdef __cplusplus
 }
